% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrs-package.R
\docType{package}
\name{ccrs-package}
\alias{ccrs-package}
\title{Correcting and Clustering preference data in the presence of response style bias.}
\description{
Corrects and clusters response-style-biased data.
}
\references{
Takagishi, M., Velden, M. van de and Yadohisa, H. (2019). Clustering preference data in the presence of response style bias, to appear in British Journal of Mathematical and Statistical Psychology.
}
\author{
Mariko Takagishi
}
\keyword{package}
