% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.X2F.R
\name{convert.X2F}
\alias{convert.X2F}
\title{Convert data matrix to rank-ordered boundary data}
\usage{
convert.X2F(X,q=q)
}
\arguments{
\item{X}{An n by m categorical data matrix.}

\item{q}{An integer indicating the maximum rating.}
}
\value{
An n by q-1 scaled rank-ordered boundary data.
}
\description{
Converts data matrix to rank-ordered boundary data.
}
