% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRSdata.R
\name{transformRSdata}
\alias{transformRSdata}
\title{Transform data by the estimated response function}
\usage{
transformRSdata(X,Beta=Beta,Mmat.q=Mmat.q)
}
\arguments{
\item{X}{An n by m categorical data matrix.}

\item{Beta}{An n by q-1 matrix of coefficiets for response functions.}

\item{Mmat.q}{A q by 3+1 matrix of I-spline basis functions, evaluated at the midpoints between boundaries.}
}
\value{
Returns a list with the following elements.
\item{\code{Y.hat}}{An n by m matrix of corrected data matrix.}
\item{\code{MB}}{An n by q matrix of values of response functions evaluated at the midpoint between boundaries.}
}
\description{
Transforms data matrix by estimated response functions.
}
